Fixed RandomSensing run issue, new issues though
So I changed the original path to use the path variable instead so that we can both run the code without repeatedly changing one line. There was an issue with the handle sense method but I fixed that. Now I have an issue with choosing moves.

Right now, chatgpt or claude or whatever you used didn't even use the available moves list lol. It just picked whatever move on the board stockfish thought was the best. I'm trying to fix it though.