Best version so far (uses both entropy and stockfish)
So the only file you need to look at is ImprovedAgent.py, that's the best version so far. What I implemented:

Constant STATE_LIMIT is at 100000, we only limit to about 100000 now it seems to work best.

If our piece was captured we sense there since that seems to be a good sensing move according to the replays.
If there is only 1 possible state we use stockfish to predict the opponents best move and sense there.
Otherwise, we use naive entropy to sense the area of most uncertainty (i got help from aidan for that method)
Using the combination of these approaches seemed to work really well for keeping the board states low so we rocking with it. I also think including both in the doc will really secure our marks.

I realised that choose moves seems to be very unoptimal, when I replayed games I noticed that there were instances where we sensed our king was in imminent danger and we still played some other random move. So it seems that choosing moves right now sucks, probably because we don't weigh the moves by their aggression or defense or something like that. But I added some code that is supposed to try defend the king if in some of our potential states we believe it will be killed. I don't feel like adding more but if something were to be done it would be to prioritize bishop moves because I noticed trout kept killing us with these longshot bishop moves that we couldn't sense.
I also removed the depth limit because I noticed that it caused our bot to make smarter moves and our number of states generally stays low so we can handle it.

Overall it looks like it beats trout quite consistently, like maybe 60-70% of the time, only because I can't fix the issue where the states become 0, but if we add the knight pushes we can bolster that. I also haven't had time to test the new code for defending the king, so just comment that out if you see weird errors happening. Timeout errors also happen quite frequently for some reason idk.